rename: App.tsx를 각 컴포넌트별로 분리
