+++ react-02-product-cart-ui/src/App.tsx
@@ -1,103 +1,5 @@
-function ProductCategoryRow({ category }: { category: string }) {
-  return (
-    <tr>
-      <th colSpan={2}>{category}</th>
-    </tr>
-  );
-}
-
-interface ProductCategoryRowProps {
-  product: Product;
-}
-
-function ProductRow({ product }: ProductCategoryRowProps) {
-  const name = product.stocked ? (
-    product.name
-  ) : (
-    <span style={{ color: "red" }}>{product.name}</span>
-  );
-
-  return (
-    <tr>
-      <td>{name}</td>
-      <td>{product.price}</td>
-    </tr>
-  );
-}
-
-interface ProductTableProps {
-  products: Product[];
-}
-
-import { JSX } from "react";
-function ProductTable({ products }: ProductTableProps) {
-  const rows: JSX.Element[] = [];
-  let lastCategory: string | null = null;
-
-  products.forEach((product) => {
-    if (product.category !== lastCategory) {
-      rows.push(
-        <ProductCategoryRow
-          category={product.category}
-          key={product.category}
-        />
-      );
-    }
-    rows.push(<ProductRow product={product} key={product.name} />);
-    lastCategory = product.category;
-  });
-
-  return (
-    <table>
-      <thead>
-        <tr>
-          <th>Name</th>
-          <th>Price</th>
-        </tr>
-      </thead>
-      <tbody>{rows}</tbody>
-    </table>
-  );
-}
-
-function SearchBar() {
-  return (
-    <form>
-      <input type="text" placeholder="Search..." />
-      <label>
-        <input type="checkbox" /> Only show products in stock
-      </label>
-    </form>
-  );
-}
-
-interface Product {
-  category: string;
-  price: string;
-  stocked: boolean;
-  name: string;
-}
-interface FilterableProductTableProps {
-  products: Product[];
-}
-
-function FilterableProductTable({ products }: FilterableProductTableProps) {
-  return (
-    <div>
-      <SearchBar />
-      <ProductTable products={products} />
-    </div>
-  );
-}
-
-const PRODUCTS = [
-  { category: "Fruits", price: "$1", stocked: true, name: "Apple" },
-  { category: "Fruits", price: "$1", stocked: true, name: "Dragonfruit" },
-  { category: "Fruits", price: "$2", stocked: false, name: "Passionfruit" },
-  { category: "Vegetables", price: "$2", stocked: true, name: "Spinach" },
-  { category: "Vegetables", price: "$4", stocked: false, name: "Pumpkin" },
-  { category: "Vegetables", price: "$1", stocked: true, name: "Peas" },
-];
+import { FilterableProductTable } from "./components/FilterableProductTable";
+import { PRODUCTS } from "./data/products";
 
 export default function App() {
   return <FilterableProductTable products={PRODUCTS} />;
