+++ react-02-product-cart-ui/src/components/ProductCategoryRow.tsx
@@ -1,0 +1,11 @@
+interface ProductCategoryRowProps {
+  category: string;
+}
+
+export function ProductCategoryRow({ category }: ProductCategoryRowProps) {
+  return (
+    <tr>
+      <th colSpan={2}>{category}</th>
+    </tr>
+  );
+}
