+++ react-02-product-cart-ui/src/components/ProductRow.tsx
@@ -1,0 +1,20 @@
+import { Product } from "../types/Product";
+
+interface ProductRowProps {
+  product: Product;
+}
+
+export function ProductRow({ product }: ProductRowProps) {
+  const name = product.stocked ? (
+    product.name
+  ) : (
+    <span style={{ color: "red" }}>{product.name}</span>
+  );
+
+  return (
+    <tr>
+      <td>{name}</td>
+      <td>{product.price}</td>
+    </tr>
+  );
+}
