+++ react-02-product-cart-ui/src/components/ProductTable.tsx
@@ -1,0 +1,38 @@
+import { JSX } from "react";
+import { Product } from "../types/Product";
+import { ProductCategoryRow } from "./ProductCategoryRow";
+import { ProductRow } from "./ProductRow";
+
+interface ProductTableProps {
+  products: Product[];
+}
+
+export function ProductTable({ products }: ProductTableProps) {
+  const rows: JSX.Element[] = [];
+  let lastCategory: string | null = null;
+
+  products.forEach((product) => {
+    if (product.category !== lastCategory) {
+      rows.push(
+        <ProductCategoryRow
+          category={product.category}
+          key={product.category}
+        />
+      );
+    }
+    rows.push(<ProductRow product={product} key={product.name} />);
+    lastCategory = product.category;
+  });
+
+  return (
+    <table>
+      <thead>
+        <tr>
+          <th>Name</th>
+          <th>Price</th>
+        </tr>
+      </thead>
+      <tbody>{rows}</tbody>
+    </table>
+  );
+}
