+++ react-02-product-cart-ui/src/components/SearchBar.tsx
@@ -1,0 +1,10 @@
+export function SearchBar() {
+  return (
+    <form>
+      <input type="text" placeholder="Search..." />
+      <label>
+        <input type="checkbox" /> Only show products in stock
+      </label>
+    </form>
+  );
+}
