+++ react-02-product-cart-ui/src/types/Product.ts
@@ -1,0 +1,6 @@
+export interface Product {
+  category: string;
+  price: string;
+  stocked: boolean;
+  name: string;
+}
